pyproject.toml
[tool.poetry]
name = "zeno-evals"
version = "0.1.1"
description = "Visualize OpenAI evals with Zeno"
authors = ["Alex Cabrera <alex.cabrera@gmail.com>"]
readme = "README.md"
packages = [{ include = "zeno_evals" }]

[tool.poetry.dependencies]
python = ">=3.8.1,<=3.11"
fire = "^0.5.0"
zenoml = "^0.4.4"
pandas = "^1.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
zeno-evals = "zeno_evals.main:cli"
